+++ src/frontend/src/pages/part/pricing/PriceBreakPanel.tsx
@@ -16,6 +16,7 @@ import useTable from '@lib/hooks/UseTable';
 import type { ApiFormFieldSet } from '@lib/types/Forms';
 import type { TableColumn } from '@lib/types/Tables';
 import { tooltipFormatter } from '../../../components/charts/tooltipFormatter';
+import { InvenTreeTable } from '../../../components/tables/InvenTreeTable';
 import { formatCurrency } from '../../../defaults/formatters';
 import {
   useCreateApiFormModal,
@@ -23,7 +24,6 @@ import {
   useEditApiFormModal
 } from '../../../hooks/UseForm';
 import { useUserState } from '../../../states/UserState';
-import { InvenTreeTable } from '../../../tables/InvenTreeTable';
 import { NoPricingData } from './PricingPanel';
 
 export default function PriceBreakPanel({